fix: Add `--install-python=false` to install invocation.
The installer may attempt to install a Python package on MacOS,
if the versions do not line up with the minimum version blessed
by the script. To do this, it attempts to run sudo.

This disables the installation with an install flag.

The flag is present in all versions of gcloud since 352.0.0.

Fixes #82.